Relocated the level data function to it's own route file
